+++ recipes/joindomain.rb
@@ -1,20 +1,20 @@
 #
-# Cookbook Name:: pbis-nativex
+# Cookbook Name:: pbis-simplyadrian
 # Recipe:: joindomain.rb
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
 # If we're in EC2, then need to dynamically determine the the OU based on region
-include_recipe 'pbis-nativex::dynamic_ou' if node['cloud']['provider'] == 'ec2'
+include_recipe 'pbis-simplyadrian::dynamic_ou' if node['cloud']['provider'] == 'ec2'
 
 ad = Chef::EncryptedDataBagItem.load("credentials", "pbis")
 
 bash "pbis join domain" do
   code lazy { <<-EOH
-  domainjoin-cli join --ou "#{node['pbis-nativex']['oupath']}" "#{node['pbis-nativex']['domain_name']}" #{ad['ad_username']} #{ad['ad_password']}
+  domainjoin-cli join --ou "#{node['pbis-simplyadrian']['oupath']}" "#{node['pbis-simplyadrian']['domain_name']}" #{ad['ad_username']} #{ad['ad_password']}
   EOH
   }
   only_if "/opt/pbis/bin/get-status |grep \"Status:        Unknown\";"