+++ recipes/updatedns.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: pbis-nativex
+# Cookbook Name:: pbis-simplyadrian
 # Recipe:: updatedns.rb
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #