+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-pbis-nativex CHANGELOG
+pbis-simplyadrian CHANGELOG
 ======================
 
-This file is used to list changes made in each version of the pbis-nativex cookbook.
+This file is used to list changes made in each version of the pbis-simplyadrian cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of pbis-nativex
+- [Adrian Herrera] - Initial release of pbis-simplyadrian
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.